Finds hex logos for given packages

Finds and downloads the best possible image that represents a hex logo for a given list of package names.

Usage

find_hex(
  pkg_names,
  output = NULL,
  repo = NULL,
  skip_known_logos = FALSE,
  overwrite = FALSE
)

Arguments

pkg_names

A single string or a vector of strings with the names of packages to find hexes for.

output

the output path where to store the found logos

repo

The repo to use to find the package logo. If none is provided, it will try to find one based on CRAN meta information. Useful when we want to minimise API calls to CRAN.

skip_known_logos

Some logos are sourced from known repos or urls. If set to TRUE, those locations will be skipped. Useful if you are getting outdated versions of logos, but slower. Defaults to FALSE.

overwrite

If a logo with the package name already exists in the output folder, should it be overwritten. Defaults to FALSE.

Value

No return, called for side effects.